feat: Allow late joining of games
Enables players to join games in progress. If the game is full, an NPC will be replaced by the new player, prioritizing replacing an Explorer NPC, then a DM NPC.

Additionally, UI elements for class selection and player stats are now conditionally displayed based on whether the player is an Explorer, not a DM, and has confirmed their class.